Added UITextFieldDelegate; Created FollowersListVC
Added a UITextFieldDelegate extension to the SearchVC and a target to the callToActionButton that listens to username submissions and then pushes the FollowersListVC onto the stack. Listens for whenever the user either inputs a username and selects the text field return button, or inputs a username, taps outside the text field view to dismiss it and then taps the call to action button "Get Followers". The username entered gets passed as data to the FollowersListVC.