Created Custom Reusable Alert
GUAlertVC is a custom alert button that is a UIView() that acts as a container for a custom GUTitleLabel, GUBodyLabel, and GUButton to display an error message to the user. Added a UIViewController+Ext file to handle extensions to the UIViewController. In this case added a presentGUAlertOnMainThread function that takes an alertTitle, message, and buttonTitle to customize and present an alert on any view controller.