Refactored NetworkManage and FollowersListVC; Created GUError enum
GUError enumeration contains a String raw value to store specific error messages and conforms to the Error protocol to be used with the Result enumeration. Refactored the NetworkManager getFollowers escaping closure to use Result<[Follower], GUError> to clean up completed() calls with simply a .success and .failure call. Refactored FollowerListsVC to use the captured value in the result reference to then perform actions based on success and failure cases.