FollowersListVC stores [Follower]; Added Diffable Data Source
When the FollowersListVC view is loaded it will try to get a list of followers from the NetworkManager. On a successful request the view will set it's own variable to store the list of Followers and update the current data snapshot on a background thread before applying the change. It will configure the variable dataSource that contains the diffable data source for the collection view. It does this by calling configureDataSource() to maintain a dynamic UICollectionViewDiffableDataSource<Section, Follower> such that it creates a new cell based on the FollowerCell custom cell view each time there is new data. When each cell is created, each followers username is set and their respective custom cell returned to the dataSource to be updated on the snapshot and displayed.