Added Pagination to Load More Followers in FollowersListVC
Extended FollowersListVC with UICollectionViewDelegate to implement scrollViewDidEndDragging. Inside scrollViewDidEndDragging, did some basic math to calculate when the user reaches the end of their scroll (end of page) so it can determine if they have more followers to load, and if so to make a network request to load in the next page of followers. In order to do this needed to create a new class variable page to increment the page of loaded users between network calls of getFollowers(username: String, page: Int)