Added Loading View for Network Requests in UIViewController Extension File
showLoading() will create a UIView, add it to the view and animate a transparent background being displayed, at the same time it will also create a UIActivityIndicatorView and add it to the new UIView and animate the activity such that it will resemble a loading screen when making network requests. dismissLoadingView() will remove the containerView along with the activityIndicator from the superView and set it nil to dismiss it.