Added a UserInfoVC to display User Info; Added getUserInfo Network Request; Refactored NetworkManager
Added a UserInfoVC that will be created and displayed when a follower cell in the collection view of the FollowerListVC is tapped on. This presents a modal view that has a swipe dismiss and done dismiss actions. Created a getUserInfo network request in the NetworkManager to test we can successfully pull a user's information and pass it to the UserInfoVC when tapping a follower cell. Refactored the NetworkManager since the getUserInfo and getFollowers network request have duplicate code. Did this by creating a generic function that will decode JSON data, as long as the data passed conforms to decodable.