Populated UserInfoVC UI with a Custom Child VC GUUserInfoHeaderVC; Created Constants Utility File
The UserInfoVC will now create a header container UIView that is pinned to the top of the screen. Once the User clicks on one of there searched followers, the network manager will get the User data and then add the child vc GUUserInfoHeaderVC to the container UIView. The GUUserInfoHeaderVC will use the passed follower data to populate the UI with there information, as well as download their avatar image to display. Created a constants utility file to store the string of the SFSymbols I would like to use.