Added a date label to the UserInfoVC; Added String and Date extensions
Added String and Date extensions to take the user.createdAt property String, convert it into a Date object, and then convert that Date back into a string formatted to the MMM yyyy date format. That string is then concatenated to "GitHub Since" to makeup the text of the new GUBodyLabel dateLabel that is anchored to the bottom of the itemViewTwo UIView.