Added Button Functionality
Setup delegate protocols in UserInfoVC and FollowersListVC. Added delegate property in GUItemInfoVC parent class along with a target to configure the action button, but left declaration and setting of the delegate and action button functionality to the subclasses GUFollowerItemVC and GURepoItemVC. The GitHub Profile button is pressed the GURepoItemVC will call didTapGitHubProfile on the delegate and pass the user. This will then prompt the delegate in UserInfoVC to call present the safari VC that was implemented in UIViewController+Ext. When the Get Followers button is pressed the GUFollowerItemVC will call didTapGetFollowers on the delegate and pass the user. This will then prompt the delegate in UserInfoVC to call the delegate property it has, which references followersListVC and call didRequestFollowers passing the username of the user to then reset the state of the FollowersListVC screen and then set the new user by calling getFollowers() again. Shortly afterwards it will dismiss the old VC to present the new followers list.