Created a PersistenceManager; Added Saving and Removing Favorites Functionality
Created an Add Button in FollowersListVC that will add the current searched user to the favorites screen if they are not already there. This will be managed by the new PersistanceManager, that contains a instance reference to UserDefaults for storing data persistently across different app sessions. When the add button is press the PersistenceManager will call updateWith to retrieve the current list of favorites (decoded from user defaults), check if the user is being added or removed from favorites, and then it will update the list accordingly and save it back to UserDefaults via the functions of the PersistenceManager.