Created Custom Favorite Cell; Set up TableView in FavoritesListVC
FavoritesListVC now contains a UITableView() that gets populated with the users list of favorite followers. Whenever the screen will appear, the table view will refresh with the current list of favorite users. If the user has no favorites, an empty state will be shown, otherwise it will reload the table view with the new favorites. The user can delete favorites using the swipe gesture on this screen, which will not only remove it from the view, but also from UserDefaults via the PersistenceManager. When a favorite is selected it will then create a new FollowersListVC screen and present it with the followers list of that favorite user. The favorites table view cells will display the username and avatar image of each favorite.